move FileDownloader and ConnectivityManager to a more specific packages
